refactor(subagents): extract stream parsing into subagent-stream.ts
Move the subagent stream-json parsing out of manager.ts into a new
src/agent/subagents/subagent-stream.ts: processStreamEvent,
parseResultFromStdout, the per-event handlers, and the ExecutionState they
thread. Pure move, no behavior change.

SubagentResult (returned by parseResultFromStdout, used by executeSubagent/
spawnSubagent) moves to the subagents index so both sides share it without a
cycle; ExecutionState lives in the new module and manager imports it back.

This brings manager.ts to 873 lines — under the 1,000-line ceiling — so its
size-baseline entry is removed entirely.
